customBasePrice, setHiddenPrice started
customBasePrice works: adds top and bottom prices together for a total
price display on arrival to the product. Still needs to have support for
discounted price on two pieces together.
setHiddenPrice will add individual prices to hidden elements in each
form on product. These will be used to show initial total price (in
customBasePrice) and to change the prod page display price based on
which items (top and/or bottom, or none) are selected.