Limited Time Offers works with list instead of json
Changed so that it will just make a product list w/ the limited time
offers list.
Each item will have a countdown, and if the end date has passed for an
item it will be removed from the list.
A default still needs to be created to show in the even that all
promotion dates have ended.
This commit has all the dirty work in it and will have dumps and
comments cleaned up in the next commit.