Product lists and results lists square image check
An attrib will be added to signify that the product image is a square.
This will be used mainly with the separates products.
If the attrib is present the data-bind width will be used to for the
height as well, if not then the data-bind height will be used.
Is turned off until the attrib is created.