product list anchor added for SEO / inline test link
Changed the div that the app-click was on to an anchor so that tlc could
set the href and allow inline product pages to be crawled.
Since it has the app-click event on it, the default action of the anchor
is intercepted and the modal will still be shown.
Also added a temporary test link for dev on the inline page.