201411 Elastic categories / filter search work
store_filter:
Added filter type to execElasticSearch that will account for items
having more than one flex edit field for color/type/etc.
This is done by getting a count of the available fields, then itterating
through them all to see if they are populated and adding them to the
filter if so.
Editing/structuing the fields may be necessary so that when one is
checked, others do not show.

bikin-sets.json, newinit:
Added custom filter options for size. Since there are two types of
sizes, add hasTextSize=true, or hasNumSize=true to page json to show the
corresponding size options.