201411 prod list formating / carousel 4 / is_app
Added styling for product lists and tlc to hide SB (bottoms) items.
Changed accessories carousel (carousel4) on homepage to ul which fixed
formatting and made it valid html (was a div w/ li's inside).
Added is_app to universal filters to prevent items not intended for the
app from showing.